requirements.txt
asgiref==3.11.0
Django==5.2.10
django-prometheus==2.4.1
djangorestframework==3.16.1
prometheus_client==0.24.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
ruff==0.14.13
sqlparse==0.5.5
